Move `build_aux` to an inherent method on `BuildManager`
